# openapi.yml
openapi: 3.0.3
info:
  title: Adapt REST API
  description: Public REST API for the Adapt chat platform
  license:
    name: AGPL-3.0
  version: 0.1.0
servers:
- url: https://api.adapt.chat
  description: Production
paths:
...
  /guilds/{guild_id}/roles:
    get:
      tags:
      - roles
      summary: Get All Roles
      description: |-
        Fetches information for all roles in the guild with the given ID.
        You must be a member of the guild.
      operationId: get_roles
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: Array of role objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You are not a member of the guild
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
    post:
      tags:
      - roles
      summary: Create Role
      description: |-
        Creates a role in the guild with the given guild ID. You must have the `MANAGE_ROLES` permission
        to create roles.
      operationId: create_role
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRolePayload'
        required: true
      responses:
        '201':
          description: Role object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: |-
            You are forbidden from creating the roles. This can be because:
            * You are not a member of the guild.
            * You do not have the `MANAGE_ROLES` permission in the guild.
            * You are creating a role that allows or denies permissions you do not have.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
    patch:
      tags:
      - roles
      summary: Edit Role Positions
      description: |-
        Modifies the positions of all roles in the guild with the given ID. You must have the
        ``MANAGE_ROLES`` permission.
      operationId: edit_role_positions
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
                minimum: 0
        required: true
      responses:
        '200':
          description: Modified role positions
        '400':
          description: "Invalid list of role IDs. Your array must contain all roles in the guild exactly\n                once, including roles above your top role but excluding the default role.\n            "
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "You are forbidden from reordering roles. This can be because:\n* You are not a member of the guild.\n* You do not have the `MANAGE_ROLES` permission in the guild.\n* A role higher than or equal to your top role had its position changed.\n            "
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guild or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
  /guilds/{guild_id}/roles/{role_id}:
    get:
      tags:
      - roles
      summary: Get Role
      description: |-
        Fetches information for the role with the given ID in the given guild. You must be a member of
        the guild the role belongs to.
      operationId: get_role
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: Role object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You are not a member of the guild
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guild or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
    delete:
      tags:
      - roles
      summary: Delete Role
      description: |-
        Deletes the role with the given ID in the given guild. You must have the `MANAGE_ROLES`
        permission, the role cannot be managed, and the role must be lower than your highest role.
      operationId: delete_role
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '204':
          description: Role deleted
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: |-
            You are forbidden from deleting the role. This can be because:
            * You are not a member of the guild.
            * You do not have the `MANAGE_ROLES` permission in the guild.
            * The role is managed.
            * The role is higher than your highest role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
    patch:
      tags:
      - roles
      summary: Edit Role
      description: |-
        Modifies the role with the given ID in the given guild. You must have the `MANAGE_ROLES`
        permission, and the role must be lower than your highest role.
      operationId: edit_role
      parameters:
      - name: guild_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRolePayload'
        required: true
      responses:
        '200':
          description: Modified role object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: |
            You are forbidden from editing the role. This can be because:
            * You are not a member of the guild.
            * You do not have the `MANAGE_ROLES` permission in the guild.
            * You are editing a role that allows or denies permissions you do not have.
            * You are editing a role that is higher than your highest role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guild or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - token: []
...
components:
  schemas:
...
    CreateRolePayload:
      type: object
      description: Payload sent to create a new role in a guild.
      required:
      - name
      properties:
        color:
          allOf:
          - $ref: '#/components/schemas/ExtendedColor'
          nullable: true
        hoisted:
          type: boolean
          description: Whether the role should be hoisted.
        icon:
          type: string
          description: |-
            The icon of the role represented as a
            [Data URI scheme](https://en.wikipedia.org/wiki/Data_URI_scheme), if any.
          nullable: true
        mentionable:
          type: boolean
          description: Whether the role should be mentionable by anyone.
        name:
          type: string
          description: The name of the role.
        permissions:
          $ref: '#/components/schemas/PermissionPair'
        position:
          type: integer
          format: int32
          description: |-
            The position the role should be in. Must be at least (and defaults to) ``1`` and at most
            the position of your top role (unless you are owner).
          minimum: 0
...
    EditRolePayload:
      type: object
      description: Payload sent to edit a role.
      properties:
        color:
          allOf:
          - $ref: '#/components/schemas/ExtendedColor'
          nullable: true
        hoisted:
          type: boolean
          description: Whether the role should be hoisted.
          nullable: true
        icon:
          type: string
          format: byte
          description: |-
            The new icon of the role. Explicitly setting this to `None` will clear the icon.
            This should be a [Data URI scheme](https://en.wikipedia.org/wiki/Data_URI_scheme).
          nullable: true
        mentionable:
          type: boolean
          description: Whether the role should be mentionable by anyone.
          nullable: true
        name:
          type: string
          description: The new name of the role, if any.
          nullable: true
        permissions:
          allOf:
          - $ref: '#/components/schemas/PermissionPair'
          nullable: true
...
    ExtendedColor:
      oneOf:
      - type: object
        description: A solid color.
        required:
        - color
        - type
        properties:
          color:
            type: integer
            format: int32
            description: The color of the solid color.
            minimum: 0
          type:
            type: string
            enum:
            - solid
      - allOf:
        - $ref: '#/components/schemas/Gradient'
        - type: object
          required:
          - type
          properties:
            type:
              type: string
              enum:
              - gradient
      description: |-
        A color that can either be solid or a linear gradient. Individual colors are specified as
        integers between 0 and 16777215.
      discriminator:
        propertyName: type
    Gradient:
      type: object
      description: |-
        A variation of an extended color that represents a linear gradient. Note that gradients are
        strictly linear and are provided in this format to allow for better consistency and ease of
        implementation in clients.
      required:
      - angle
      - stops
      properties:
        angle:
          type: number
          format: float
          description: The angle of the gradient, in radians.
        stops:
          type: array
          items:
            $ref: '#/components/schemas/GradientStop'
          description: The color stops of the gradient.
    GradientStop:
      type: object
      description: A single color stop in a linear gradient.
      required:
      - position
      - color
      properties:
        color:
          type: integer
          format: int32
          description: The color of the stop.
          minimum: 0
        position:
          type: number
          format: float
          description: The position of the stop in the gradient, between 0 and 1.
...
    PermissionPair:
      type: object
      description: |-
        Represents a pair of permissions, one representing allowed permissions and the other
        representing denied permissions. This is so that any permission that is represented as
        "neutral" where it is neither allowed/denied remains easily overwritten by lower
        roles or members.
      required:
      - allow
      - deny
      properties:
        allow:
          $ref: '#/components/schemas/Permissions'
        deny:
          $ref: '#/components/schemas/Permissions'
    Permissions:
      type: integer
      format: int64
...
    Role:
      type: object
      description: A role in a guild.
      required:
      - id
      - guild_id
      - name
      - permissions
      - position
      - flags
      properties:
        color:
          allOf:
          - $ref: '#/components/schemas/ExtendedColor'
          nullable: true
        flags:
          $ref: '#/components/schemas/RoleFlags'
        guild_id:
          type: integer
          format: int64
          description: The ID of the guild this role belongs to.
          minimum: 0
        icon:
          type: string
          description: The URL of the role's icon, if any, displayed alongside member names in chat.
          nullable: true
        id:
          type: integer
          format: int64
          description: The snowflake ID of the role.
          minimum: 0
        name:
          type: string
          description: The name of the role.
        permissions:
          $ref: '#/components/schemas/PermissionPair'
        position:
          type: integer
          format: int32
          description: |-
            The position of this role in the role hierarchy. The lower the number, the lower the role.
            The default role always has a position of 0.

            The backend will try its best to keep all role positions unique, but on the event two
            collide due to something such as a data race, then the true position of these roles will
            not be predictable, and will likely be in the order of model creation.
          minimum: 0
    RoleFlags:
      type: integer
      format: int32
...
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: Authorization